Get  the  current user and check filter resources according to the person Implements only on the reservation still have to implement in the remains domain models

+++ flightreservation/src/main/java/edu/miu/cs/flightreservation/controller/ReservationController.java
@@ -2,6 +2,7 @@ package edu.miu.cs.flightreservation.controller;
 
 import edu.miu.cs.flightreservation.Util.payload.request.ReservationRequest;
 import edu.miu.cs.flightreservation.Util.payload.request.ReservationStatusRequest;
+import edu.miu.cs.flightreservation.Util.security.jwt.JwtUtils;
 import edu.miu.cs.flightreservation.model.*;
 import edu.miu.cs.flightreservation.service.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,8 +12,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -31,26 +35,34 @@ public class ReservationController {
     private TicketServiceImp ticketService;
 
     @GetMapping()
-    public ResponseEntity<List<Reservation>> getReservationsByPage(){
-        List<Reservation> reservations = reservationService.findAll();
-        return new ResponseEntity<>(reservations, HttpStatus.OK);
+    public ResponseEntity<List<Reservation>> getReservationsByPage(HttpServletRequest httpServletRequest){
+        Person auth = personService.getPersonFromToken(httpServletRequest);
+        List<Reservation> reservations = new ArrayList<Reservation>();
+        if(auth != null){
+            reservations = reservationService.findAllByPersonId(auth.getId());
+            return new ResponseEntity<>(reservations, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(reservations, HttpStatus.NO_CONTENT);
     }
 
+    //TODO this must not  be availlable for passenger
     @GetMapping(params = "page=true")
     public ResponseEntity<Page<Reservation>> getReservationsByPage(Pageable pageable){
         Page<Reservation> reservations = reservationService.findAll(pageable);
         return new ResponseEntity<>(reservations, HttpStatus.OK);
     }
 
+    //TODO ROLE BASED AUTHORIZATION
     @PostMapping()
     @Transactional(Transactional.TxType.REQUIRES_NEW)
-    public ResponseEntity<Reservation> createReservation(@RequestBody @Valid ReservationRequest reservation){
+    public ResponseEntity<Reservation> createReservation(HttpServletRequest httpServletRequest,
+                                                         @RequestBody @Valid ReservationRequest reservation){
+        Person auth = personService.getPersonFromToken(httpServletRequest);
         try{
             long[] flights = reservation.getFlights();
-            //if(!flightService.exists(flights))
-                //return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
-
-            //TODO check if all flights are in the database
+            if(!flightService.exists(flights))
+                return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+            //TODO Discuss about it we need to check if the user exist we must just get the passenger_id
             Person person = new Person();
             person.setFirstName(reservation.getFirstName());
             person.setLastName(reservation.getLastName());
@@ -61,11 +73,10 @@ public class ReservationController {
             //TODO generate the encrypted password
             person.setPassword("password");
             Person _person = personService.save(person);
-
             Reservation _reservation = new Reservation();
             _reservation.setDepartureDate(reservation.getDepartureDate());
-            //TODO get the current authenticated user and pass it
-            _reservation.setCreatedBy(null);
+            _reservation.setCreatedBy(auth);
+            _reservation.setPassenger(_person);
             _reservation.setArrivalDate(reservation.getArrivalDate());
             _reservation.setDepartureDate(reservation.getDepartureDate());
             _reservation.setArrivalPlace(reservation.getArrivalPlace());
@@ -107,19 +118,28 @@ public class ReservationController {
     }*/
 
     @PutMapping("/{id}")
-    public ResponseEntity<Reservation> updateReservationStatus(@PathVariable("id") Long id, @RequestBody ReservationStatusRequest reservation){
+    public ResponseEntity<Reservation> updateReservationStatus(HttpServletRequest httpServletRequest,
+                                                               @PathVariable("id") Long id,
+                                                               @RequestBody ReservationStatusRequest reservation){
+        Person auth = personService.getPersonFromToken(httpServletRequest);
         Reservation _reservation = reservationService.findById(id);
         if(_reservation != null){
+            if(!_reservation.getCreatedBy().equals(auth) && !_reservation.getPassenger().equals(auth))
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
             reservationService.updateStatus(_reservation, reservation.getStatus());
                 return new ResponseEntity<>(reservationService.save(_reservation), HttpStatus.OK);
         }else
             return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
     }
 
+    //TODO ROLE BASED AUTHORIZATION PASSENGER CAN DO THIS
     @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteReservation(@PathVariable("id") Long id){
+    public ResponseEntity<HttpStatus> deleteReservation(HttpServletRequest httpServletRequest, @PathVariable("id") Long id){
+        Person auth = personService.getPersonFromToken(httpServletRequest);
         Reservation _reservation = reservationService.findById(id);
         if(_reservation != null){
+            if(_reservation.getPassenger().equals(auth))
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
             reservationService.delete(_reservation);
             return new ResponseEntity<>(HttpStatus.OK);
         }else
